Adjustments and refactoring
